requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
chardet==4.0.0
discord.py==2.0.1
idna==3.2
multidict==5.1.0
pycodestyle==2.7.0
python-dotenv==0.18.0
toml==0.10.2
typing-extensions==3.10.0.0
yarl==1.6.3
psycopg==3.0.14
psycopg2-binary==2.9.3
asyncpg
Pillow
requires.io
pycord